MixMeet.Auth/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
python-dotenv==1.0.0
redis==5.0.1
python-jose[cryptography]==3.3.0 
requests==2.31.0